+++ lib/src/build_tracker.dart
@@ -8,28 +8,28 @@ import 'tracking_build_owner.dart';
 part 'build_tracker.freezed.dart';
 
 @freezed
-class BuildEntry with _$BuildEntry {
-  factory BuildEntry({
+class RebuildDirtyWidget with _$RebuildDirtyWidget {
+  factory RebuildDirtyWidget({
     required int timestamp,
     required String widget,
-  }) = _BuildEntry;
+  }) = _RebuildDirtyWidget;
 }
 
 @freezed
-class ScheduleBuildEntry with _$ScheduleBuildEntry {
-  factory ScheduleBuildEntry({
+class ScheduleBuildFor with _$ScheduleBuildFor {
+  factory ScheduleBuildFor({
     required int timestamp,
     required String widget,
     required BuiltList<String> stack,
-  }) = _ScheduleBuildEntry;
+  }) = _ScheduleBuildFor;
 }
 
 @freezed
 class BuildFrame with _$BuildFrame {
   factory BuildFrame({
     required int number,
-    required BuiltList<BuildEntry> built,
-    required BuiltList<ScheduleBuildEntry> buildScheduled,
+    required BuiltList<RebuildDirtyWidget> rebuildDirtyWidgets,
+    required BuiltList<ScheduleBuildFor> schedueBuildFors,
   }) = _BuildFrame;
 }
 
@@ -42,8 +42,8 @@ class BuildTracker {
   BuildTracker({
     this.onBuildFrame,
     this.printBuildFrame = true,
-    this.printBuildFrameIncludeBuilt = true,
-    this.printBuildFrameIncludeBuildScheduled = true,
+    this.printBuildFrameIncludeRebuildDirtyWidget = true,
+    this.printBuildFrameIncludeScheduleBuildFor = true,
     bool enabled = true,
   }) {
     this.enabled = enabled;
@@ -55,14 +55,14 @@ class BuildTracker {
   bool printBuildFrame;
 
   ///
-  /// Print every widget that was build.
+  /// Print every widget that was built.
   ///
-  bool printBuildFrameIncludeBuilt;
+  bool printBuildFrameIncludeRebuildDirtyWidget;
 
   ///
-  /// Print every widget that was marked as dirty.
+  /// Print every widget for which [BuildOwner.scheduleBuildFor] was called with stack traces (build roots).
   ///
-  bool printBuildFrameIncludeBuildScheduled;
+  bool printBuildFrameIncludeScheduleBuildFor;
 
   ///
   /// Called after every frame with `BuildFrame` information collected during the frame build.
@@ -77,10 +77,16 @@ class BuildTracker {
     }
     _enabled = value;
     if (value) {
-      assert(debugOnRebuildDirtyWidget == null,
-          "`debugOnRebuildDirtyWidget` already in use ($debugOnRebuildDirtyWidget)");
-      assert(debugOnScheduleBuildFor == null,
-          "`debugOnScheduleBuildFor` already in use ($debugOnScheduleBuildFor)");
+      assert(
+        debugOnRebuildDirtyWidget == null,
+        "`debugOnRebuildDirtyWidget` already in use ($debugOnRebuildDirtyWidget)",
+      );
+
+      assert(
+        debugOnScheduleBuildFor == null,
+        "`debugOnScheduleBuildFor` already in use ($debugOnScheduleBuildFor)",
+      );
+
       assert(
         WidgetsBinding.instance is TrackingBuildOwnerWidgetsBindingMixin,
         "`TrackingBuildOwnerWidgetsBindingMixin` is required (${WidgetsBinding.instance})",
@@ -96,37 +102,46 @@ class BuildTracker {
   }
 
   void _onDebugOnRebuildDirtyWidget(Element e, bool builtOnce) {
-    _buildList.add(BuildEntry(
+    _buildList.add(RebuildDirtyWidget(
         timestamp: DateTime.now().millisecondsSinceEpoch,
         widget: e.debugGetCreatorChain(10)));
   }
 
   void _onDebugOnScheduleBuildFor(Element e) {
     var chain = Chain.current();
-    final setStateIndex = chain.traces.first.frames.lastIndexWhere((_) => {
-          'State.setState',
-          'Element.markNeedsBuild',
-          'HookState.setState'
-        }.contains(_.member));
+    final setStateIndex = chain.traces.first.frames.lastIndexWhere(
+      (_) => {
+        'State.setState',
+        'Element.markNeedsBuild',
+        'HookState.setState',
+      }.contains(_.member),
+    );
     if (setStateIndex > 0) {
       chain = Chain.current(setStateIndex);
     }
-    _buildScheduleList.add(ScheduleBuildEntry(
+    _buildScheduleList.add(
+      ScheduleBuildFor(
         timestamp: DateTime.now().millisecondsSinceEpoch,
         widget: e.debugGetCreatorChain(10),
         stack: chain.terse.traces
             .expand((t) => t.frames)
-            .map((f) => '${f.member},${f.location}')
-            .toBuiltList()));
+            .map(
+              (f) => '${f.member},${f.location}',
+            )
+            .toBuiltList(),
+      ),
+    );
   }
 
   void _frameCallback(Duration _) {
     _frameCallbackScheduled = false;
 
     final frame = BuildFrame(
-        number: _number++,
-        built: BuiltList(_buildList),
-        buildScheduled: BuiltList(_buildScheduleList));
+      number: _number++,
+      rebuildDirtyWidgets: BuiltList(_buildList),
+      schedueBuildFors: BuiltList(_buildScheduleList),
+    );
+
     _buildList.clear();
     _buildScheduleList.clear();
 
@@ -149,38 +164,44 @@ class BuildTracker {
     debugPrint('# BuildTracker frame #${frame.number}');
     debugPrint('');
 
-    if (printBuildFrameIncludeBuilt && frame.built.isNotEmpty) {
+    if (printBuildFrameIncludeRebuildDirtyWidget &&
+        frame.rebuildDirtyWidgets.isNotEmpty) {
       debugPrint('## Widgets that were built');
       debugPrint('');
-      for (final e in frame.built) {
+      for (final e in frame.rebuildDirtyWidgets) {
         debugPrint('- `${e.widget}`');
       }
       debugPrint('');
     }
 
-    if (printBuildFrameIncludeBuildScheduled &&
-        frame.buildScheduled.isNotEmpty) {
+    if (printBuildFrameIncludeScheduleBuildFor &&
+        frame.schedueBuildFors.isNotEmpty) {
       debugPrint('## Widgets that were marked dirty (build roots)');
       debugPrint('');
-      for (final e in frame.buildScheduled) {
-        final stackHash = e.stack.hashCode.toRadixString(16);
-        final printedIn =
-            _printedStacks.putIfAbsent(e.stack, () => frame.number);
+      for (final e in frame.schedueBuildFors) {
+        final stack = e.stack;
+
+        final stackHash = stack.hashCode.toRadixString(16);
+        _printedStacksCounts[stack] = (_printedStacksCounts[stack] ?? 0) + 1;
+
+        final printedIn = _printedStacksFirstFrame[stack];
         debugPrint('### ${e.widget}:');
         debugPrint('');
-        if (printedIn != frame.number) {
+        if (printedIn != null) {
           debugPrint(
               'Stack trace #$stackHash observed in frame #$printedIn for the first time');
         } else {
+          _printedStacksFirstFrame[stack] = frame.number;
+
           debugPrint('Stack trace #$stackHash:');
           debugPrint('```');
-          for (final frame in e.stack.map((_) => _.split(','))) {
+          for (final frame in stack.map((_) => _.split(','))) {
             final member = frame[0];
             final location = frame[1];
             final isCore = {
               'dart:',
               'package:flutter/',
-              'package:flutter_test/'
+              'package:flutter_test/',
             }.any((_) => location.startsWith(_));
             debugPrint(
                 '${isCore ? ' ' : '*'} ${member.padRight(40)} $location');
@@ -195,13 +216,63 @@ class BuildTracker {
     debugPrint('');
   }
 
+  ///
+  /// Print markdown-formatted top [BuildOwner.scheduleBuildFor] stack traces.
+  ///
+  void printTopScheduleBuildForStacks({
+    int count = 10,
+    bool reset = true,
+  }) {
+    final top = _printedStacksCounts.entries.toList()
+      ..sortBy<num>((_) => -_.value)
+      ..take(count);
+
+    debugPrint('## Top $count `scheduleBuildFor` stack traces (build roots)');
+    debugPrint('');
+    for (final e in top) {
+      final stack = e.key;
+
+      final stackHash = stack.hashCode.toRadixString(16);
+
+      debugPrint('### ${e.value} times:');
+      debugPrint('');
+
+      debugPrint('Stack trace #$stackHash:');
+      debugPrint('```');
+      for (final frame in stack.map((_) => _.split(','))) {
+        final member = frame[0];
+        final location = frame[1];
+        final isCore = {
+          'dart:',
+          'package:flutter/',
+          'package:flutter_test/',
+        }.any((_) => location.startsWith(_));
+        debugPrint('${isCore ? ' ' : '*'} ${member.padRight(40)} $location');
+      }
+      debugPrint('```');
+      debugPrint('');
+    }
+
+    if (reset) {
+      resetScheduledBuildForStacksCounts();
+    }
+  }
+
+  ///
+  /// Reset [BuildOwner.scheduleBuildFor] stack traces counts.
+  ///
+  void resetScheduledBuildForStacksCounts() {
+    _printedStacksCounts.clear();
+  }
+
   var _enabled = false;
   var _number = 1;
 
   var _frameCallbackScheduled = false;
 
-  final _buildList = <BuildEntry>[];
-  final _buildScheduleList = <ScheduleBuildEntry>[];
+  final _buildList = <RebuildDirtyWidget>[];
+  final _buildScheduleList = <ScheduleBuildFor>[];
 
-  final _printedStacks = <BuiltList<String>, int>{};
+  final _printedStacksFirstFrame = <BuiltList<String>, int>{};
+  final _printedStacksCounts = <BuiltList<String>, int>{};
 }