+++ lib/src/build_tracker.freezed.dart
@@ -13,11 +13,11 @@ final _privateConstructorUsedError = UnsupportedError(
     'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
 /// @nodoc
-class _$BuildEntryTearOff {
-  const _$BuildEntryTearOff();
+class _$RebuildDirtyWidgetTearOff {
+  const _$RebuildDirtyWidgetTearOff();
 
-  _BuildEntry call({required int timestamp, required String widget}) {
-    return _BuildEntry(
+  _RebuildDirtyWidget call({required int timestamp, required String widget}) {
+    return _RebuildDirtyWidget(
       timestamp: timestamp,
       widget: widget,
     );
@@ -25,33 +25,34 @@ class _$BuildEntryTearOff {
 }
 
 /// @nodoc
-const $BuildEntry = _$BuildEntryTearOff();
+const $RebuildDirtyWidget = _$RebuildDirtyWidgetTearOff();
 
 /// @nodoc
-mixin _$BuildEntry {
+mixin _$RebuildDirtyWidget {
   int get timestamp => throw _privateConstructorUsedError;
   String get widget => throw _privateConstructorUsedError;
 
   @JsonKey(ignore: true)
-  $BuildEntryCopyWith<BuildEntry> get copyWith =>
+  $RebuildDirtyWidgetCopyWith<RebuildDirtyWidget> get copyWith =>
       throw _privateConstructorUsedError;
 }
 
 /// @nodoc
-abstract class $BuildEntryCopyWith<$Res> {
-  factory $BuildEntryCopyWith(
-          BuildEntry value, $Res Function(BuildEntry) then) =
-      _$BuildEntryCopyWithImpl<$Res>;
+abstract class $RebuildDirtyWidgetCopyWith<$Res> {
+  factory $RebuildDirtyWidgetCopyWith(
+          RebuildDirtyWidget value, $Res Function(RebuildDirtyWidget) then) =
+      _$RebuildDirtyWidgetCopyWithImpl<$Res>;
   $Res call({int timestamp, String widget});
 }
 
 /// @nodoc
-class _$BuildEntryCopyWithImpl<$Res> implements $BuildEntryCopyWith<$Res> {
-  _$BuildEntryCopyWithImpl(this._value, this._then);
+class _$RebuildDirtyWidgetCopyWithImpl<$Res>
+    implements $RebuildDirtyWidgetCopyWith<$Res> {
+  _$RebuildDirtyWidgetCopyWithImpl(this._value, this._then);
 
-  final BuildEntry _value;
+  final RebuildDirtyWidget _value;
   // ignore: unused_field
-  final $Res Function(BuildEntry) _then;
+  final $Res Function(RebuildDirtyWidget) _then;
 
   @override
   $Res call({
@@ -72,30 +73,32 @@ class _$BuildEntryCopyWithImpl<$Res> implements $BuildEntryCopyWith<$Res> {
 }
 
 /// @nodoc
-abstract class _$BuildEntryCopyWith<$Res> implements $BuildEntryCopyWith<$Res> {
-  factory _$BuildEntryCopyWith(
-          _BuildEntry value, $Res Function(_BuildEntry) then) =
-      __$BuildEntryCopyWithImpl<$Res>;
+abstract class _$RebuildDirtyWidgetCopyWith<$Res>
+    implements $RebuildDirtyWidgetCopyWith<$Res> {
+  factory _$RebuildDirtyWidgetCopyWith(
+          _RebuildDirtyWidget value, $Res Function(_RebuildDirtyWidget) then) =
+      __$RebuildDirtyWidgetCopyWithImpl<$Res>;
   @override
   $Res call({int timestamp, String widget});
 }
 
 /// @nodoc
-class __$BuildEntryCopyWithImpl<$Res> extends _$BuildEntryCopyWithImpl<$Res>
-    implements _$BuildEntryCopyWith<$Res> {
-  __$BuildEntryCopyWithImpl(
-      _BuildEntry _value, $Res Function(_BuildEntry) _then)
-      : super(_value, (v) => _then(v as _BuildEntry));
+class __$RebuildDirtyWidgetCopyWithImpl<$Res>
+    extends _$RebuildDirtyWidgetCopyWithImpl<$Res>
+    implements _$RebuildDirtyWidgetCopyWith<$Res> {
+  __$RebuildDirtyWidgetCopyWithImpl(
+      _RebuildDirtyWidget _value, $Res Function(_RebuildDirtyWidget) _then)
+      : super(_value, (v) => _then(v as _RebuildDirtyWidget));
 
   @override
-  _BuildEntry get _value => super._value as _BuildEntry;
+  _RebuildDirtyWidget get _value => super._value as _RebuildDirtyWidget;
 
   @override
   $Res call({
     Object? timestamp = freezed,
     Object? widget = freezed,
   }) {
-    return _then(_BuildEntry(
+    return _then(_RebuildDirtyWidget(
       timestamp: timestamp == freezed
           ? _value.timestamp
           : timestamp // ignore: cast_nullable_to_non_nullable
@@ -110,8 +113,8 @@ class __$BuildEntryCopyWithImpl<$Res> extends _$BuildEntryCopyWithImpl<$Res>
 
 /// @nodoc
 
-class _$_BuildEntry implements _BuildEntry {
-  _$_BuildEntry({required this.timestamp, required this.widget});
+class _$_RebuildDirtyWidget implements _RebuildDirtyWidget {
+  _$_RebuildDirtyWidget({required this.timestamp, required this.widget});
 
   @override
   final int timestamp;
@@ -120,13 +123,13 @@ class _$_BuildEntry implements _BuildEntry {
 
   @override
   String toString() {
-    return 'BuildEntry(timestamp: $timestamp, widget: $widget)';
+    return 'RebuildDirtyWidget(timestamp: $timestamp, widget: $widget)';
   }
 
   @override
   bool operator ==(dynamic other) {
     return identical(this, other) ||
-        (other is _BuildEntry &&
+        (other is _RebuildDirtyWidget &&
             (identical(other.timestamp, timestamp) ||
                 const DeepCollectionEquality()
                     .equals(other.timestamp, timestamp)) &&
@@ -142,13 +145,13 @@ class _$_BuildEntry implements _BuildEntry {
 
   @JsonKey(ignore: true)
   @override
-  _$BuildEntryCopyWith<_BuildEntry> get copyWith =>
-      __$BuildEntryCopyWithImpl<_BuildEntry>(this, _$identity);
+  _$RebuildDirtyWidgetCopyWith<_RebuildDirtyWidget> get copyWith =>
+      __$RebuildDirtyWidgetCopyWithImpl<_RebuildDirtyWidget>(this, _$identity);
 }
 
-abstract class _BuildEntry implements BuildEntry {
-  factory _BuildEntry({required int timestamp, required String widget}) =
-      _$_BuildEntry;
+abstract class _RebuildDirtyWidget implements RebuildDirtyWidget {
+  factory _RebuildDirtyWidget(
+      {required int timestamp, required String widget}) = _$_RebuildDirtyWidget;
 
   @override
   int get timestamp => throw _privateConstructorUsedError;
@@ -156,19 +159,19 @@ abstract class _BuildEntry implements BuildEntry {
   String get widget => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  _$BuildEntryCopyWith<_BuildEntry> get copyWith =>
+  _$RebuildDirtyWidgetCopyWith<_RebuildDirtyWidget> get copyWith =>
       throw _privateConstructorUsedError;
 }
 
 /// @nodoc
-class _$ScheduleBuildEntryTearOff {
-  const _$ScheduleBuildEntryTearOff();
+class _$ScheduleBuildForTearOff {
+  const _$ScheduleBuildForTearOff();
 
-  _ScheduleBuildEntry call(
+  _ScheduleBuildFor call(
       {required int timestamp,
       required String widget,
       required BuiltList<String> stack}) {
-    return _ScheduleBuildEntry(
+    return _ScheduleBuildFor(
       timestamp: timestamp,
       widget: widget,
       stack: stack,
@@ -177,35 +180,35 @@ class _$ScheduleBuildEntryTearOff {
 }
 
 /// @nodoc
-const $ScheduleBuildEntry = _$ScheduleBuildEntryTearOff();
+const $ScheduleBuildFor = _$ScheduleBuildForTearOff();
 
 /// @nodoc
-mixin _$ScheduleBuildEntry {
+mixin _$ScheduleBuildFor {
   int get timestamp => throw _privateConstructorUsedError;
   String get widget => throw _privateConstructorUsedError;
   BuiltList<String> get stack => throw _privateConstructorUsedError;
 
   @JsonKey(ignore: true)
-  $ScheduleBuildEntryCopyWith<ScheduleBuildEntry> get copyWith =>
+  $ScheduleBuildForCopyWith<ScheduleBuildFor> get copyWith =>
       throw _privateConstructorUsedError;
 }
 
 /// @nodoc
-abstract class $ScheduleBuildEntryCopyWith<$Res> {
-  factory $ScheduleBuildEntryCopyWith(
-          ScheduleBuildEntry value, $Res Function(ScheduleBuildEntry) then) =
-      _$ScheduleBuildEntryCopyWithImpl<$Res>;
+abstract class $ScheduleBuildForCopyWith<$Res> {
+  factory $ScheduleBuildForCopyWith(
+          ScheduleBuildFor value, $Res Function(ScheduleBuildFor) then) =
+      _$ScheduleBuildForCopyWithImpl<$Res>;
   $Res call({int timestamp, String widget, BuiltList<String> stack});
 }
 
 /// @nodoc
-class _$ScheduleBuildEntryCopyWithImpl<$Res>
-    implements $ScheduleBuildEntryCopyWith<$Res> {
-  _$ScheduleBuildEntryCopyWithImpl(this._value, this._then);
+class _$ScheduleBuildForCopyWithImpl<$Res>
+    implements $ScheduleBuildForCopyWith<$Res> {
+  _$ScheduleBuildForCopyWithImpl(this._value, this._then);
 
-  final ScheduleBuildEntry _value;
+  final ScheduleBuildFor _value;
   // ignore: unused_field
-  final $Res Function(ScheduleBuildEntry) _then;
+  final $Res Function(ScheduleBuildFor) _then;
 
   @override
   $Res call({
@@ -231,25 +234,25 @@ class _$ScheduleBuildEntryCopyWithImpl<$Res>
 }
 
 /// @nodoc
-abstract class _$ScheduleBuildEntryCopyWith<$Res>
-    implements $ScheduleBuildEntryCopyWith<$Res> {
-  factory _$ScheduleBuildEntryCopyWith(
-          _ScheduleBuildEntry value, $Res Function(_ScheduleBuildEntry) then) =
-      __$ScheduleBuildEntryCopyWithImpl<$Res>;
+abstract class _$ScheduleBuildForCopyWith<$Res>
+    implements $ScheduleBuildForCopyWith<$Res> {
+  factory _$ScheduleBuildForCopyWith(
+          _ScheduleBuildFor value, $Res Function(_ScheduleBuildFor) then) =
+      __$ScheduleBuildForCopyWithImpl<$Res>;
   @override
   $Res call({int timestamp, String widget, BuiltList<String> stack});
 }
 
 /// @nodoc
-class __$ScheduleBuildEntryCopyWithImpl<$Res>
-    extends _$ScheduleBuildEntryCopyWithImpl<$Res>
-    implements _$ScheduleBuildEntryCopyWith<$Res> {
-  __$ScheduleBuildEntryCopyWithImpl(
-      _ScheduleBuildEntry _value, $Res Function(_ScheduleBuildEntry) _then)
-      : super(_value, (v) => _then(v as _ScheduleBuildEntry));
+class __$ScheduleBuildForCopyWithImpl<$Res>
+    extends _$ScheduleBuildForCopyWithImpl<$Res>
+    implements _$ScheduleBuildForCopyWith<$Res> {
+  __$ScheduleBuildForCopyWithImpl(
+      _ScheduleBuildFor _value, $Res Function(_ScheduleBuildFor) _then)
+      : super(_value, (v) => _then(v as _ScheduleBuildFor));
 
   @override
-  _ScheduleBuildEntry get _value => super._value as _ScheduleBuildEntry;
+  _ScheduleBuildFor get _value => super._value as _ScheduleBuildFor;
 
   @override
   $Res call({
@@ -257,7 +260,7 @@ class __$ScheduleBuildEntryCopyWithImpl<$Res>
     Object? widget = freezed,
     Object? stack = freezed,
   }) {
-    return _then(_ScheduleBuildEntry(
+    return _then(_ScheduleBuildFor(
       timestamp: timestamp == freezed
           ? _value.timestamp
           : timestamp // ignore: cast_nullable_to_non_nullable
@@ -276,8 +279,8 @@ class __$ScheduleBuildEntryCopyWithImpl<$Res>
 
 /// @nodoc
 
-class _$_ScheduleBuildEntry implements _ScheduleBuildEntry {
-  _$_ScheduleBuildEntry(
+class _$_ScheduleBuildFor implements _ScheduleBuildFor {
+  _$_ScheduleBuildFor(
       {required this.timestamp, required this.widget, required this.stack});
 
   @override
@@ -289,13 +292,13 @@ class _$_ScheduleBuildEntry implements _ScheduleBuildEntry {
 
   @override
   String toString() {
-    return 'ScheduleBuildEntry(timestamp: $timestamp, widget: $widget, stack: $stack)';
+    return 'ScheduleBuildFor(timestamp: $timestamp, widget: $widget, stack: $stack)';
   }
 
   @override
   bool operator ==(dynamic other) {
     return identical(this, other) ||
-        (other is _ScheduleBuildEntry &&
+        (other is _ScheduleBuildFor &&
             (identical(other.timestamp, timestamp) ||
                 const DeepCollectionEquality()
                     .equals(other.timestamp, timestamp)) &&
@@ -314,15 +317,15 @@ class _$_ScheduleBuildEntry implements _ScheduleBuildEntry {
 
   @JsonKey(ignore: true)
   @override
-  _$ScheduleBuildEntryCopyWith<_ScheduleBuildEntry> get copyWith =>
-      __$ScheduleBuildEntryCopyWithImpl<_ScheduleBuildEntry>(this, _$identity);
+  _$ScheduleBuildForCopyWith<_ScheduleBuildFor> get copyWith =>
+      __$ScheduleBuildForCopyWithImpl<_ScheduleBuildFor>(this, _$identity);
 }
 
-abstract class _ScheduleBuildEntry implements ScheduleBuildEntry {
-  factory _ScheduleBuildEntry(
+abstract class _ScheduleBuildFor implements ScheduleBuildFor {
+  factory _ScheduleBuildFor(
       {required int timestamp,
       required String widget,
-      required BuiltList<String> stack}) = _$_ScheduleBuildEntry;
+      required BuiltList<String> stack}) = _$_ScheduleBuildFor;
 
   @override
   int get timestamp => throw _privateConstructorUsedError;
@@ -332,7 +335,7 @@ abstract class _ScheduleBuildEntry implements ScheduleBuildEntry {
   BuiltList<String> get stack => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  _$ScheduleBuildEntryCopyWith<_ScheduleBuildEntry> get copyWith =>
+  _$ScheduleBuildForCopyWith<_ScheduleBuildFor> get copyWith =>
       throw _privateConstructorUsedError;
 }
 
@@ -342,12 +345,12 @@ class _$BuildFrameTearOff {
 
   _BuildFrame call(
       {required int number,
-      required BuiltList<BuildEntry> built,
-      required BuiltList<ScheduleBuildEntry> buildScheduled}) {
+      required BuiltList<RebuildDirtyWidget> rebuildDirtyWidgets,
+      required BuiltList<ScheduleBuildFor> schedueBuildFors}) {
     return _BuildFrame(
       number: number,
-      built: built,
-      buildScheduled: buildScheduled,
+      rebuildDirtyWidgets: rebuildDirtyWidgets,
+      schedueBuildFors: schedueBuildFors,
     );
   }
 }
@@ -358,8 +361,9 @@ const $BuildFrame = _$BuildFrameTearOff();
 /// @nodoc
 mixin _$BuildFrame {
   int get number => throw _privateConstructorUsedError;
-  BuiltList<BuildEntry> get built => throw _privateConstructorUsedError;
-  BuiltList<ScheduleBuildEntry> get buildScheduled =>
+  BuiltList<RebuildDirtyWidget> get rebuildDirtyWidgets =>
+      throw _privateConstructorUsedError;
+  BuiltList<ScheduleBuildFor> get schedueBuildFors =>
       throw _privateConstructorUsedError;
 
   @JsonKey(ignore: true)
@@ -374,8 +378,8 @@ abstract class $BuildFrameCopyWith<$Res> {
       _$BuildFrameCopyWithImpl<$Res>;
   $Res call(
       {int number,
-      BuiltList<BuildEntry> built,
-      BuiltList<ScheduleBuildEntry> buildScheduled});
+      BuiltList<RebuildDirtyWidget> rebuildDirtyWidgets,
+      BuiltList<ScheduleBuildFor> schedueBuildFors});
 }
 
 /// @nodoc
@@ -389,22 +393,22 @@ class _$BuildFrameCopyWithImpl<$Res> implements $BuildFrameCopyWith<$Res> {
   @override
   $Res call({
     Object? number = freezed,
-    Object? built = freezed,
-    Object? buildScheduled = freezed,
+    Object? rebuildDirtyWidgets = freezed,
+    Object? schedueBuildFors = freezed,
   }) {
     return _then(_value.copyWith(
       number: number == freezed
           ? _value.number
           : number // ignore: cast_nullable_to_non_nullable
               as int,
-      built: built == freezed
-          ? _value.built
-          : built // ignore: cast_nullable_to_non_nullable
-              as BuiltList<BuildEntry>,
-      buildScheduled: buildScheduled == freezed
-          ? _value.buildScheduled
-          : buildScheduled // ignore: cast_nullable_to_non_nullable
-              as BuiltList<ScheduleBuildEntry>,
+      rebuildDirtyWidgets: rebuildDirtyWidgets == freezed
+          ? _value.rebuildDirtyWidgets
+          : rebuildDirtyWidgets // ignore: cast_nullable_to_non_nullable
+              as BuiltList<RebuildDirtyWidget>,
+      schedueBuildFors: schedueBuildFors == freezed
+          ? _value.schedueBuildFors
+          : schedueBuildFors // ignore: cast_nullable_to_non_nullable
+              as BuiltList<ScheduleBuildFor>,
     ));
   }
 }
@@ -417,8 +421,8 @@ abstract class _$BuildFrameCopyWith<$Res> implements $BuildFrameCopyWith<$Res> {
   @override
   $Res call(
       {int number,
-      BuiltList<BuildEntry> built,
-      BuiltList<ScheduleBuildEntry> buildScheduled});
+      BuiltList<RebuildDirtyWidget> rebuildDirtyWidgets,
+      BuiltList<ScheduleBuildFor> schedueBuildFors});
 }
 
 /// @nodoc
@@ -434,22 +438,22 @@ class __$BuildFrameCopyWithImpl<$Res> extends _$BuildFrameCopyWithImpl<$Res>
   @override
   $Res call({
     Object? number = freezed,
-    Object? built = freezed,
-    Object? buildScheduled = freezed,
+    Object? rebuildDirtyWidgets = freezed,
+    Object? schedueBuildFors = freezed,
   }) {
     return _then(_BuildFrame(
       number: number == freezed
           ? _value.number
           : number // ignore: cast_nullable_to_non_nullable
               as int,
-      built: built == freezed
-          ? _value.built
-          : built // ignore: cast_nullable_to_non_nullable
-              as BuiltList<BuildEntry>,
-      buildScheduled: buildScheduled == freezed
-          ? _value.buildScheduled
-          : buildScheduled // ignore: cast_nullable_to_non_nullable
-              as BuiltList<ScheduleBuildEntry>,
+      rebuildDirtyWidgets: rebuildDirtyWidgets == freezed
+          ? _value.rebuildDirtyWidgets
+          : rebuildDirtyWidgets // ignore: cast_nullable_to_non_nullable
+              as BuiltList<RebuildDirtyWidget>,
+      schedueBuildFors: schedueBuildFors == freezed
+          ? _value.schedueBuildFors
+          : schedueBuildFors // ignore: cast_nullable_to_non_nullable
+              as BuiltList<ScheduleBuildFor>,
     ));
   }
 }
@@ -459,19 +463,19 @@ class __$BuildFrameCopyWithImpl<$Res> extends _$BuildFrameCopyWithImpl<$Res>
 class _$_BuildFrame implements _BuildFrame {
   _$_BuildFrame(
       {required this.number,
-      required this.built,
-      required this.buildScheduled});
+      required this.rebuildDirtyWidgets,
+      required this.schedueBuildFors});
 
   @override
   final int number;
   @override
-  final BuiltList<BuildEntry> built;
+  final BuiltList<RebuildDirtyWidget> rebuildDirtyWidgets;
   @override
-  final BuiltList<ScheduleBuildEntry> buildScheduled;
+  final BuiltList<ScheduleBuildFor> schedueBuildFors;
 
   @override
   String toString() {
-    return 'BuildFrame(number: $number, built: $built, buildScheduled: $buildScheduled)';
+    return 'BuildFrame(number: $number, rebuildDirtyWidgets: $rebuildDirtyWidgets, schedueBuildFors: $schedueBuildFors)';
   }
 
   @override
@@ -480,19 +484,20 @@ class _$_BuildFrame implements _BuildFrame {
         (other is _BuildFrame &&
             (identical(other.number, number) ||
                 const DeepCollectionEquality().equals(other.number, number)) &&
-            (identical(other.built, built) ||
-                const DeepCollectionEquality().equals(other.built, built)) &&
-            (identical(other.buildScheduled, buildScheduled) ||
+            (identical(other.rebuildDirtyWidgets, rebuildDirtyWidgets) ||
                 const DeepCollectionEquality()
-                    .equals(other.buildScheduled, buildScheduled)));
+                    .equals(other.rebuildDirtyWidgets, rebuildDirtyWidgets)) &&
+            (identical(other.schedueBuildFors, schedueBuildFors) ||
+                const DeepCollectionEquality()
+                    .equals(other.schedueBuildFors, schedueBuildFors)));
   }
 
   @override
   int get hashCode =>
       runtimeType.hashCode ^
       const DeepCollectionEquality().hash(number) ^
-      const DeepCollectionEquality().hash(built) ^
-      const DeepCollectionEquality().hash(buildScheduled);
+      const DeepCollectionEquality().hash(rebuildDirtyWidgets) ^
+      const DeepCollectionEquality().hash(schedueBuildFors);
 
   @JsonKey(ignore: true)
   @override
@@ -503,15 +508,16 @@ class _$_BuildFrame implements _BuildFrame {
 abstract class _BuildFrame implements BuildFrame {
   factory _BuildFrame(
       {required int number,
-      required BuiltList<BuildEntry> built,
-      required BuiltList<ScheduleBuildEntry> buildScheduled}) = _$_BuildFrame;
+      required BuiltList<RebuildDirtyWidget> rebuildDirtyWidgets,
+      required BuiltList<ScheduleBuildFor> schedueBuildFors}) = _$_BuildFrame;
 
   @override
   int get number => throw _privateConstructorUsedError;
   @override
-  BuiltList<BuildEntry> get built => throw _privateConstructorUsedError;
+  BuiltList<RebuildDirtyWidget> get rebuildDirtyWidgets =>
+      throw _privateConstructorUsedError;
   @override
-  BuiltList<ScheduleBuildEntry> get buildScheduled =>
+  BuiltList<ScheduleBuildFor> get schedueBuildFors =>
       throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)